+++ Package.swift
@@ -4,8 +4,8 @@ import PackageDescription
 let package = Package(
   name: "rswift",
   dependencies: [
-    .package(url: "https://github.com/kylef/Commander.git", .upToNextMinor(from: "0.6.0")),
-    .package(url: "https://github.com/tomlokhorst/XcodeEdit", .upToNextMinor(from: "2.0.0"))
+    .package(url: "https://github.com/kylef/Commander.git", from: "0.8.0"),
+    .package(url: "https://github.com/tomlokhorst/XcodeEdit", from: "2.4.1")
   ],
   targets: [
     .target(