+++ Sources/RswiftCore/Generators/AggregatedStructGenerator.swift
@@ -22,7 +22,13 @@ class AggregatedStructGenerator: StructGenerator {
     let internalStructName: SwiftIdentifier = "_R"
 
     let collectedResult = subgenerators
-      .map { $0.generatedStructs(at: externalAccessLevel, prefix: qualifiedName) }
+      .compactMap {
+        let result = $0.generatedStructs(at: externalAccessLevel, prefix: qualifiedName)
+        if result.externalStruct.isEmpty { return nil }
+        if let internalStruct = result.internalStruct, internalStruct.isEmpty { return nil }
+
+        return result
+      }
       .reduce(StructGeneratorResultCollector()) { collector, result in collector.appending(result) }
 
     let externalStruct = Struct(
@@ -72,7 +78,7 @@ private struct StructGeneratorResultCollector {
   func appending(_ result: StructGenerator.Result) -> StructGeneratorResultCollector {
     return StructGeneratorResultCollector(
       externalStructs: externalStructs + [result.externalStruct],
-      internalStructs: internalStructs + [result.internalStruct].flatMap { $0 }
+      internalStructs: internalStructs + [result.internalStruct].compactMap { $0 }
     )
   }
 }