+++ Sources/RswiftCore/Generators/NibStructGenerator.swift
@@ -35,6 +35,11 @@ private let Ordinals = [
 struct NibStructGenerator: StructGenerator {
   private let nibs: [Nib]
 
+  private let instantiateParameters = [
+    Function.Parameter(name: "owner", localName: "ownerOrNil", type: Type._AnyObject.asOptional()),
+    Function.Parameter(name: "options", localName: "optionsOrNil", type: Type(module: .stdLib, name: SwiftIdentifier(rawValue: "[UINib.OptionsKey : Any]"), optional: true), defaultValue: "nil")
+  ]
+
   init(nibs: [Nib]) {
     self.nibs = nibs
   }
@@ -65,7 +70,41 @@ struct NibStructGenerator: StructGenerator {
       .map { nibVar(for: $0, at: externalAccessLevel, prefix: qualifiedName) }
     let nibFunctions: [Function] = groupedNibs
       .uniques
-      .map { nibFunc(for: $0, at: externalAccessLevel, prefix: qualifiedName) }
+      .flatMap { nib -> [Function] in
+        let qualifiedCurrentNibName = qualifiedName + SwiftIdentifier(name: nib.name)
+
+        let deprecatedFunction = Function(
+          availables: ["*, deprecated, message: \"Use UINib(resource: \(qualifiedCurrentNibName)) instead\""],
+          comments: ["`UINib(name: \"\(nib.name)\", in: bundle)`"],
+          accessModifier: externalAccessLevel,
+          isStatic: true,
+          name: SwiftIdentifier(name: nib.name),
+          generics: nil,
+          parameters: [
+            Function.Parameter(name: "_", type: Type._Void, defaultValue: "()")
+          ],
+          doesThrow: false,
+          returnType: Type._UINib,
+          body: "return UIKit.UINib(resource: \(qualifiedCurrentNibName))"
+        )
+
+        guard let firstViewInfo = nib.rootViews.first else { return [deprecatedFunction] }
+
+        let newFunction = Function(
+          availables: [],
+          comments: [],
+          accessModifier: externalAccessLevel,
+          isStatic: true,
+          name: SwiftIdentifier(name: nib.name),
+          generics: nil,
+          parameters: instantiateParameters,
+          doesThrow: false,
+          returnType: firstViewInfo.asOptional(),
+          body: "return \(qualifiedCurrentNibName).instantiate(withOwner: ownerOrNil, options: optionsOrNil)[0] as? \(firstViewInfo)"
+        )
+
+        return [deprecatedFunction, newFunction]
+      }
 
     let externalStruct = Struct(
       availables: [],
@@ -86,26 +125,6 @@ struct NibStructGenerator: StructGenerator {
     )
   }
 
-  private func nibFunc(for nib: Nib, at externalAccessLevel: AccessLevel, prefix: SwiftIdentifier) -> Function {
-    let nibName = SwiftIdentifier(name: nib.name)
-    let qualifiedName = prefix + nibName
-
-    return Function(
-      availables: [],
-      comments: ["`UINib(name: \"\(nib.name)\", in: bundle)`"],
-      accessModifier: externalAccessLevel,
-      isStatic: true,
-      name: SwiftIdentifier(name: nib.name),
-      generics: nil,
-      parameters: [
-        Function.Parameter(name: "_", type: Type._Void, defaultValue: "()")
-      ],
-      doesThrow: false,
-      returnType: Type._UINib,
-      body: "return UIKit.UINib(resource: \(qualifiedName))"
-    )
-  }
-
   private func nibVar(for nib: Nib, at externalAccessLevel: AccessLevel, prefix: SwiftIdentifier) -> Let {
     let structName = SwiftIdentifier(name: "_\(nib.name)")
     let qualifiedName = prefix + structName
@@ -121,11 +140,6 @@ struct NibStructGenerator: StructGenerator {
   }
 
   private func nibStruct(for nib: Nib, at externalAccessLevel: AccessLevel) -> Struct {
-    let instantiateParameters = [
-      Function.Parameter(name: "owner", localName: "ownerOrNil", type: Type._AnyObject.asOptional()),
-      Function.Parameter(name: "options", localName: "optionsOrNil", type: Type(module: .stdLib, name: SwiftIdentifier(rawValue: "[NSObject : AnyObject]"), optional: true), defaultValue: "nil")
-    ]
-
     let bundleLet = Let(
       comments: [],
       accessModifier: externalAccessLevel,
@@ -184,10 +198,17 @@ struct NibStructGenerator: StructGenerator {
     }
 
     // Validation
-    let validateImagesLines = Set(nib.usedImageIdentifiers)
+    let validateImagesLines = nib.usedImageIdentifiers.uniqueAndSorted()
       .map {
         "if UIKit.UIImage(named: \"\($0)\", in: R.hostingBundle, compatibleWith: nil) == nil { throw Rswift.ValidationError(description: \"[R.swift] Image named '\($0)' is used in nib '\(nib.name)', but couldn't be loaded.\") }"
-    }
+      }
+    let validateColorLines = nib.usedColorResources.uniqueAndSorted()
+      .map {
+        "if UIKit.UIColor(named: \"\($0)\") == nil { throw Rswift.ValidationError(description: \"[R.swift] Color named '\($0)' is used in storyboard '\(nib.name)', but couldn't be loaded.\") }"
+      }
+    let validateColorLinesWithAvailableIf = ["if #available(iOS 11.0, *) {"] +
+      validateColorLines.map { $0.indent(with: "  ") } +
+      ["}"]
 
     var validateFunctions: [Function] = []
     var validateImplements: [Type] = []
@@ -202,7 +223,7 @@ struct NibStructGenerator: StructGenerator {
         parameters: [],
         doesThrow: true,
         returnType: Type._Void,
-        body: validateImagesLines.joined(separator: "\n")
+        body: (validateImagesLines + validateColorLinesWithAvailableIf).joined(separator: "\n")
       )
       validateFunctions.append(validateFunction)
       validateImplements.append(Type.Validatable)