+++ Sources/RswiftCore/ResourceTypes/Xcodeproj.swift
@@ -39,25 +39,25 @@ struct Xcodeproj: WhiteListedExtensionsResourceType {
 
   func resourcePathsForTarget(_ targetName: String) throws -> [Path] {
     // Look for target in project file
-    let allTargets = projectFile.project.targets.flatMap { $0.value }
+    let allTargets = projectFile.project.targets.compactMap { $0.value }
     guard let target = allTargets.filter({ $0.name == targetName }).first else {
-      let availableTargets = allTargets.flatMap { $0.name }.joined(separator: ", ")
+      let availableTargets = allTargets.compactMap { $0.name }.joined(separator: ", ")
       throw ResourceParsingError.parsingFailed("Target '\(targetName)' not found in project file, available targets are: \(availableTargets)")
     }
 
     let resourcesFileRefs = target.buildPhases
-      .flatMap { $0.value as? PBXResourcesBuildPhase }
+      .compactMap { $0.value as? PBXResourcesBuildPhase }
       .flatMap { $0.files }
-      .flatMap { $0.value?.fileRef }
+      .compactMap { $0.value?.fileRef }
 
     let fileRefPaths = resourcesFileRefs
-      .flatMap { $0.value as? PBXFileReference }
-      .flatMap { $0.fullPath }
+      .compactMap { $0.value as? PBXFileReference }
+      .compactMap { $0.fullPath }
 
     let variantGroupPaths = resourcesFileRefs
-      .flatMap { $0.value as? PBXVariantGroup }
+      .compactMap { $0.value as? PBXVariantGroup }
       .flatMap { $0.fileRefs }
-      .flatMap { $0.value?.fullPath }
+      .compactMap { $0.value?.fullPath }
 
     return fileRefPaths + variantGroupPaths
   }