+++ Sources/rswift/Rswift.swift
@@ -10,6 +10,6 @@
 import Foundation
 
 struct Rswift {
-  static let version = "4.0.0"
+  static let version = "5.0.0.alpha.2"
   static let resourceFileName = "R.generated.swift"
 }